Skip linking XS code for pure Perl install
When `PUREPERL_ONLY=1` set, this prevents the Makefile from having rules
that compile the XS code. This is useful to have when building
dependencies for deploying to a pure Perl install such as when using
`App::FatPacker`. This removes the need to skip over files that end in
`perl -V:dlext` (e.g., `.so`, `.dll`, `.dylib`) when copying to a
deployment.

This should allow for using the `cpanm --pureperl` option so that a pure
Perl installation can be requested for all installed distributions.

Signed-off-by: Zakariyya Mughal <zaki.mughal@gmail.com>